refactor: Create src folder
