+++ src/domain/definitions/submission-status-updates-queue-manager.go
@@ -1,6 +1,6 @@
 package definitions
 
-import "github.com/upb-code-labs/tests-microservice/domain/dtos"
+import "github.com/upb-code-labs/tests-microservice/src/domain/dtos"
 
 type SubmissionStatusUpdatesQueueManager interface {
 	QueueUpdate(updateDTO *dtos.SubmissionStatusUpdateDTO) error
