+++ src/infrastructure/static_files/handler.go
@@ -8,8 +8,8 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/upb-code-labs/tests-microservice/domain/dtos"
-	"github.com/upb-code-labs/tests-microservice/infrastructure"
+	"github.com/upb-code-labs/tests-microservice/src/domain/dtos"
+	"github.com/upb-code-labs/tests-microservice/src/infrastructure"
 )
 
 type StaticFilesManager struct{}
